dung-creating a new entity class Adress, creating some tests related to this Entity, updating DataInitialiser class

+++ back-end/src/main/java/com/dnth_underdog_241/online_fashion_shopping/repository/RoleRepository.java
@@ -9,7 +9,7 @@ import java.util.Optional;
 
 
 @Repository
-public interface RoleRepository extends CrudRepository<Role, Integer>
+public interface RoleRepository extends CrudRepository<Role, Long>
 {
     public Optional<Role> findByName(String name);
 };
